feat[Ibexa 5.0] Rename classes prefixed with `Ez*` to `Ibexa*`
