+++ Core/FieldHandler/IbexaDateAndTime.php
@@ -5,7 +5,7 @@ namespace Kaliop\IbexaMigrationBundle\Core\FieldHandler;
 use Kaliop\IbexaMigrationBundle\API\FieldValueConverterInterface;
 use Kaliop\IbexaMigrationBundle\API\FieldDefinitionConverterInterface;
 
-class EzDateAndTime extends AbstractFieldHandler implements FieldValueConverterInterface, FieldDefinitionConverterInterface
+class IbexaDateAndTime extends AbstractFieldHandler implements FieldValueConverterInterface, FieldDefinitionConverterInterface
 {
     /**
      * @param int|string|null $fieldValue use a timestamp; if as string, prepend it with @ sign
