+++ Core/FieldHandler/IbexaImage.php
@@ -5,7 +5,7 @@ namespace Kaliop\IbexaMigrationBundle\Core\FieldHandler;
 use Ibexa\Core\FieldType\Image\Value as ImageValue;
 use Kaliop\IbexaMigrationBundle\API\FieldValueConverterInterface;
 
-class EzImage extends FileFieldHandler implements FieldValueConverterInterface
+class IbexaImage extends FileFieldHandler implements FieldValueConverterInterface
 {
     /**
      * Creates a value object to use as the field value when setting an image field type.
