+++ Core/FieldHandler/IbexaMedia.php
@@ -5,7 +5,7 @@ namespace Kaliop\IbexaMigrationBundle\Core\FieldHandler;
 use Ibexa\Core\FieldType\Media\Value as MediaValue;
 use Kaliop\IbexaMigrationBundle\API\FieldValueConverterInterface;
 
-class EzMedia extends FileFieldHandler implements FieldValueConverterInterface
+class IbexaMedia extends FileFieldHandler implements FieldValueConverterInterface
 {
     /**
      * Creates a value object to use as the field value when setting a media field type.
