+++ WHATSNEW.md
@@ -1,6 +1,7 @@
 Version: 2.0.0
 ==============
 
+* Changed: Rename classes prefixed with `Ez*` to `Ibexa*`
 * Changed: Update namespace `\EzSystems\EzPlatformMatrixFieldtype\` to `\Ibexa\FieldTypeMatrix\`
 * Changed: Update \Kaliop\IbexaMigrationBundle\Core\Process\Process to Symfony 7
 * Changed: Update Doctrine\DBAL calls to Symfony 7
